% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Set and check parameter settings}
\usage{
create_config(..., data = NULL)
}
\arguments{
\item{data}{an optional list of data items as returned by
\code{outbreaker_data}; if provided, this allows for further checks of
the outbreaker setings.}
}
\description{
This function defines settings.  It takes a list of named
items as input, performs various checks, set defaults where arguments are
missing, and return a correct list of settings. If no input is given, it
returns the default settings.
}
\examples{
## see default settings
create_config()

## change defaults
create_config(move_alpha = FALSE, n_iter = 2e5, sample_every = 1000)



}
\seealso{
\code{\link{outbreaker_data}} to check and process data for outbreaker
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
