% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_clustering.R
\name{pre_clustering}
\alias{pre_clustering}
\title{Pre cluster cases in groups according using the genotypes and the arbitrary
thresholds gamma (spatial) and delta (temporal).}
\usage{
pre_clustering(data, config)
}
\arguments{
\item{data}{A list of data items as returned by \code{outbreaker_data}, or
arguments passed to this function.}

\item{config}{A list of settings as returned by \code{create_config}, or
arguments passed to this function.}
}
\value{
A list containing two components \code{$data} and
\code{$config}. \code{data} data items as returned by \code{outbreaker_data}.
\code{config} is a list of settings as returned by \code{create_config}. \cr \cr
}
\description{
This function updates the clusters and the  initial tree in the lists data 
and config
}
