% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{toy_outbreak}
\alias{toy_outbreak}
\title{Simulated outbreak}
\format{An object of class \code{list} of length 4.}
\usage{
toy_outbreak
}
\description{
We generated a data containing 500 cases, from simulated outbreaks across the US,
used to illustrate \code{measlesoutbreaker}. This list contains the following:
}
\details{
\itemize{

\item \code{$cases}: A data table summarising the epidemiological features of the
500 cases. It contains the State, onset date,  genotype,  county, age group,
import status, cluster and generation of the cases.

\item \code{$distance}: A matrix indicating the distance between each county

\item \code{$population}: A vector indicating the population per county

\item \code{$age_contact}: A matrix indicating the number of contacts between 
age groups

}
}
\examples{
data("toy_outbreak")
names(toy_outbreak)
toy_outbreak

}
\author{
Alexis Robert \email{alexis.robert@lshtm.ac.uk}
}
\keyword{datasets}
