% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{toy_outbreak_long}
\alias{toy_outbreak_long}
\title{Simulated outbreaks}
\format{An object of class \code{list} of length 3.}
\usage{
toy_outbreak_long
}
\description{
We generated two datasets used to illustrate \code{o2geosocial}. 
The first one (toy_outbreak_long) contains 1000 cases, from simulated outbreaks 
nationwide between 2010 and 2017. The list contains the following:
}
\details{
\itemize{

\item \code{$cases}: A data table summarising the epidemiological features of the
1000 cases. It contains the State, onset date,  genotype,  county, age group,
import status, cluster and generation of the cases.

\item \code{$dt_regions}: A data table containing the ID, population, longitude 
and latitude of each region. Should be used to compute the distance matrix, using 
the package geosphere.

\item \code{$age_contact}: A matrix indicating the number of contacts between 
age groups

}
}
\examples{
data("toy_outbreak_long")
names(toy_outbreak_long)
toy_outbreak_long

}
\author{
Alexis Robert \email{alexis.robert@lshtm.ac.uk}
}
\keyword{datasets}
