% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaker.R
\name{outbreaker}
\alias{outbreaker}
\title{outbreaker: main function for reconstructing disease outbreaks}
\usage{
outbreaker(
  data = outbreaker_data(),
  config = create_config(),
  priors = custom_priors(),
  likelihoods = custom_likelihoods(),
  moves = custom_moves()
)
}
\arguments{
\item{data}{a list of named items containing input data as returned by
\code{\link{outbreaker_data}}}

\item{config}{a set of settings as returned by \code{\link{create_config}}}

\item{priors}{a set of log-prior functions as returned by
\code{\link{custom_priors}}}

\item{likelihoods}{a set of log-likelihood functions as returned by
\code{\link{custom_likelihoods}}}

\item{moves}{a set of movement functions as returned by
\code{\link{custom_moves}}}
}
\value{
A data frame of \code{n_iter / sample_every} rows (as defined in the function\code{create_config()}). For each row, the data frame contains:
\itemize{

\item \code{post}: The posterior value of the transmission tree at this iteration.

\item \code{like}: The likelihood value of the transmission tree at this iteration.

\item \code{post}: The posterior value of the transmission tree at this iteration.

\item \code{a}: The estimate of the spatial parameter \code{a} at this iteration,

\item \code{a}: The estimate of the spatial parameter \code{b} at this iteration,

\item \code{pi}: The estimate of the conditional report ratio \code{pi} at this iteration,

\item \code{alpha_1 to alpha_N}: The infector of each case at this iteration.

\item \code{t_inf_1 to t_inf_N}: The infection date of each case at this iteration.

\item \code{kappa_1 to kappa_N}: The number of generation between each case and their infector at this iteration.

}
}
\description{
The function \code{outbreaker} is the main function of the package. It runs
processes various inputs (data, configuration settings, custom priors,
likelihoods and movement functions) and explores the space of plausible
transmission trees of a densely sampled outbreaks.\cr
}
\examples{

## get data
data(toy_outbreak_short)

\dontrun{
## run outbreaker
dt_cases <- toy_outbreak_short$cases
dt_cases <- dt_cases[order(dt_cases$Date), ]
dt_regions <- toy_outbreak_short$dt_regions
all_dist <- geosphere::distGeo(matrix(c(rep(dt_regions$long, nrow(dt_regions)), 
                                        rep(dt_regions$lat, nrow(dt_regions))), 
                                      ncol = 2), 
                               matrix(c(rep(dt_regions$long, each = nrow(dt_regions)), 
                                        rep(dt_regions$lat, each = nrow(dt_regions))),
                                      ncol = 2))

dist_mat <- matrix(all_dist/1000, nrow = nrow(dt_regions))
pop_vect <- dt_regions$population
names(pop_vect) <- rownames(dist_mat) <- colnames(dist_mat) <- dt_regions$region

data <- outbreaker_data(dates = dt_cases$Date, age_group = dt_cases$age_group,
                        region = dt_cases$Cens_tract, population = pop_vect, 
                        distance = dist_mat, a_dens = toy_outbreak_short$age_contact,
                        f_dens = dgamma(x = 1:300, scale = 0.43, shape = 27),
                        w_dens = dnorm(x = 1:300, mean = 11.7, sd = 2.0))
out <- outbreaker(data = data, config = list(n_iter = 200, sample_every = 5,
                                             n_iter_import = 100, sample_every_import = 5,
                                             gamma = 100, delta = 30, burnin = 20))
plot(out)


}
}
\seealso{
\code{\link{outbreaker_data}} to process input data, and
    \code{\link{create_config}} to process/set up parameters

\itemize{

\item \code{\link{outbreaker_data}}: function to process input data

\item \code{\link{create_config}}: function to create default and customise
configuration settings

\item \code{\link{custom_priors}}: function to specify customised prior
functions

\item \code{\link{custom_likelihoods}}: function to specify customised likelihoods
functions

\item \code{\link{custom_moves}}: function to create default and customise movement
functions

}
}
\author{
Initial version by Thibaut Jombart, rewritten by Alexis Robert (\email{alexis.robert@lshtm.ac.uk})
}
